+++ app/src/main/java/org/fitchfamily/android/symphony/MusicService.java
@@ -867,6 +867,16 @@ public class MusicService extends Service implements
             String trackAlbum = songToPlay.getAlbum();
             String trackArtist = songToPlay.getArtist();
 
+            // If we are in random album mode, then swap artist and album so that the Grom
+            // USB-3 adaptor will display the album (symphony, Broadway show, etc.) title
+            // in the "DISC" field on the Prius multi-function display instead of its default
+            // artist information. See https://github.com/n76/Symphony/issues/5
+
+            if (shuffle == PLAY_RANDOM_ALBUM) {
+                String t = trackAlbum;
+                trackAlbum = trackArtist;
+                trackArtist = t;
+            }
 
             Bitmap albumArt = BitmapFactory.decodeResource(getResources(),
                     R.drawable.violin_icon); //replace with medias albumArt